+++ Wemotely/Controllers/JobsTableViewController.swift
@@ -1,4 +1,5 @@
 import UIKit
+import FeedKit
 import RealmSwift
 
 protocol JobsTableViewControllerDelegate: class {
@@ -7,6 +8,8 @@ protocol JobsTableViewControllerDelegate: class {
 
 class JobsTableViewController: UITableViewController {
     let realm = RealmProvider.realm()
+
+    internal let refresher = UIRefreshControl()
 
     weak var delegate: DashboardTableViewController?
     var row: Row?
@@ -22,6 +25,45 @@ class JobsTableViewController: UITableViewController {
         tableView.accessibilityIdentifier = "jobsTableView"
 
         didEdit = false
+
+        tableView.refreshControl = refresher
+        // Add Last Updated
+        refresher.attributedTitle = NSAttributedString(string: "Pull to refresh jobs")
+        refresher.addTarget(self, action: #selector(refreshJobs(_:)), for: .valueChanged)
+    }
+
+    @objc private func refreshJobs(_ sender: Any) {
+        logger.info("-> Refreshing Job Data")
+
+        var accounts: Results<Account>? {
+            // Only Refresh Accounts that we want
+            if let accountUUID = row?.accountUUID {
+                return Account.allByUUID(provider: realm, uuid: accountUUID)
+            } else {
+                return Account.activeSorted(provider: realm)
+            }
+        }
+
+        // Move to helper function
+        for account in accounts! {
+            var feed: RSSFeed!
+            let feedService = FeedService(account: account)
+
+            let feedURL = URLProvider(key: account.urlKey!).url()
+
+            feedService.parser(url: feedURL)?.parseAsync(queue: DispatchQueue.global(qos: .userInitiated)) { (result) in
+                feed = result.rssFeed!
+
+                DispatchQueue.main.async {
+                    feedService.save(realm: self.realm, feed: feed)
+
+                    logger.info("-> Refreshed Job Data: \(String(describing: feed.items?.count))")
+
+                    self.refresher.endRefreshing()
+                    self.tableView.reloadData()
+                }
+            }
+        }
     }
 
     override func viewWillDisappear(_ animated: Bool) {
